% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_with_and.R
\name{collapse_with_and}
\alias{collapse_with_and}
\title{Collapse values into a grammatical list}
\usage{
collapse_with_and(x, sep = ", ", connector = "and", serial_comma = TRUE)
}
\arguments{
\item{x}{A vector of values.}
\item{sep}{A character string used to separate the elements.}
\item{connector}{A character string to place between the final two elements, such as \code{and} or \code{or}. Use \code{NULL} or \code{""} for no connector.}
\item{serial_comma}{Logical; whether to include \code{sep} before \code{connector} in lists of three or more elements. Defaults to \code{TRUE}.}
}
\value{
A single character string.
}
\description{
Collapse values into a grammatical list
}
\examples{
collapse_with_and(c(1, 2, 3))
collapse_with_and(c("apples", "bananas", "pears"), connector = "or")
collapse_with_and(c("apples", "bananas", "pears"), serial_comma = FALSE)
collapse_with_and(c("apples", "bananas"), connector = NULL)
}
