% Generated by roxygen2 (4.0.0): do not edit by hand
\name{createResultsTemplate}
\alias{createResultsTemplate}
\title{Creates a result template database table}
\usage{
createResultsTemplate(table)
}
\arguments{
\item{table}{name of the database table}
}
\description{
Builds a SQL query to create a table designed to hold one year worth of results
}
\examples{
createResultsTemplate()
}
\keyword{create}
\keyword{database}
\keyword{function}
\keyword{template}

