% Generated by roxygen2 (4.0.0): do not edit by hand
\name{createTable}
\alias{createTable}
\title{Creates a database table}
\usage{
createTable(table, colNames, colType)
}
\arguments{
\item{table}{name of the database table}

\item{colNames}{the names of the columns in the table to be created}

\item{colType}{defining characteristics of each column}
}
\description{
Builds a SQL query to inset a list of values into a database table
}
\examples{
createTable()
}
\keyword{create}
\keyword{database}
\keyword{function}

