% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fillTable}
\alias{fillTable}
\title{Fill a database table}
\usage{
fillTable(table, columns, valMat)
}
\arguments{
\item{table}{name of the database table}

\item{columns}{the names of the columns to be inserted into}

\item{a}{matrix of values to be inserted}
}
\description{
Builds a SQL query to inset a list of values into a database table
}
\examples{
fillTable()
}
\keyword{database}
\keyword{fill}
\keyword{function}

