% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decade.R
\name{get_decade}
\alias{get_decade}
\title{Get the decade for a date}
\usage{
get_decade(x)
}
\arguments{
\item{x}{A date or date-time vector.}
}
\value{
A numeric vector giving the first year of the decade for each value
in \code{x}.
}
\description{
Get the decade for a date
}
\examples{
get_decade(as.Date(c("1974-06-01", "1989-01-01", "2026-09-05")))
}
