% Generated by roxygen2 (4.0.0): do not edit by hand
\name{importFlows}
\alias{importFlows}
\title{Import Flow Cytometry Files (.csv)}
\usage{
importFlows(fileNames)
}
\arguments{
\item{fileNames}{The .csv files you wish to import}
}
\description{
Imports one or many .csv files containing flow cytometry data
}
\examples{
importFlows()
}
\keyword{flow}
\keyword{import}

