% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Transform log-odds into probabilities}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A number or vector.}
}
\description{
Transform log-odds into probabilities
}
\examples{
inv_logit(c(-2, -1, 0, 1, 2))
}
