% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_hex.R
\name{mix_hex}
\alias{mix_hex}
\title{Mix two hexadecimal colours}
\usage{
mix_hex(col1, col2, alpha = 0.5)
}
\arguments{
\item{col1}{A hexadecimal colour string.}
\item{col2}{A hexadecimal colour string.}
\item{alpha}{Numeric; the mixing proportion. \code{0} returns \code{col1}, and \code{1} returns \code{col2}.}
}
\value{
A hexadecimal colour string.
}
\description{
Mix two hexadecimal colours
}
\examples{
mix_hex("#E69F00", "#56B4E9")
mix_hex("#E69F00", "#56B4E9", alpha = 0.25)
}
