% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_word.R
\name{number_to_word}
\alias{number_to_word}
\title{Convert numbers from zero to ten into words}
\usage{
number_to_word(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector, with whole numbers from zero to ten converted
to words and all other values converted to character strings.
}
\description{
Convert numbers from zero to ten into words
}
\examples{
number_to_word(c(0, 5, 10, 11))
}
