% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_sd.R
\name{pop_sd}
\alias{pop_sd}
\title{Calculate a population standard deviation}
\usage{
pop_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
\item{na.rm}{Logical; should missing values be removed?}
}
\value{
A numeric value giving the population standard deviation of \code{x}.
}
\description{
Calculate a population standard deviation
}
\examples{
pop_sd(c(1, 2, 3, 4))
}
