% Generated by roxygen2 (4.0.0): do not edit by hand
\name{readInfileResults}
\alias{readInfileResults}
\title{Reads in a simulation result file}
\usage{
readInfileResults(table, file)
}
\arguments{
\item{table}{name of the database table}

\item{file}{file to read results from}
}
\description{
Builds a SQL query to read in one year's worth of results
}
\examples{
readInfileResults()
}
\keyword{database}
\keyword{read}
\keyword{results}

