% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_list_of_numbers.R
\\name{nice_list_of_numbers}
\\alias{nice_list_of_numbers}
\\title{Format a vector of numbers as a grammatical list}
\\usage{
nice_list_of_numbers(x, digits = NULL)
}
\\arguments{
\\item{x}{A numeric vector.}
\\item{digits}{An optional number of decimal places to round to before formatting.}
}
\\value{
A single character string.
}
\\description{
Format a vector of numbers as a grammatical list
}
\\examples{
nice_list_of_numbers(c(1, 2, 3))
nice_list_of_numbers(c(1.234, 5.678), digits = 2)
}
